requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
statsmodels>=0.14.0
PyYAML>=6.0
